Funpay-Steam-Guard/requirements.txt
python-dotenv>=1.0.0
hmac
base64
time
logging
